none in thumb.jpg fixed

+++ load_img.sh
@@ -5,16 +5,23 @@ set -euo pipefail
 DIR="$1"
 IFS=$'\t' read -ra f
 
-curl -so "$DIR/thumb.jpg" "${f[6]}"
-curl -so "$DIR/cover_tmp.jpg" "${f[7]}"
+if [ "${f[6]}" != "Null" ]; then
+    curl -so "$DIR/thumb.jpg" "${f[6]}"
+else
+    echo "no thumb.jpg"
+fi
 
-s=${f[8]}; t=${f[9]}; m=$s; (( t < s )) && m=$t
+if [ "${f[7]}" != "Null" ]; then
+    curl -so "$DIR/cover_tmp.jpg" "${f[7]}"
 
-convert "$DIR/cover_tmp.jpg" -gravity center -crop ${m}x${m}+0+0 +repage "$DIR/cover.jpg" 2>/dev/null
+    s=${f[8]}; t=${f[9]}; m=$s; (( t < s )) && m=$t
 
-rm "$DIR/cover_tmp.jpg"
+    convert "$DIR/cover_tmp.jpg" -gravity center -crop ${m}x${m}+0+0 +repage "$DIR/cover.jpg" 2>/dev/null
 
-
+    rm "$DIR/cover_tmp.jpg"
+else
+    echo "no cover.jpg"
+fi
 
 printf '%s\t%s\t%s\t%s\t%s\t%s\t%s\t%s\n' "${f[@]:0:6}" "$DIR/thumb.jpg" "$DIR/cover.jpg"
